% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_fastqs.R
\name{create_fq_sheet}
\alias{create_fq_sheet}
\title{Create a table with details on names of fastq files}
\usage{
create_fq_sheet(x, ext = opts_flow$get("fastq_extension"),
  format = opts_flow$get("fastq_format"), lst_patterns = list(samplename =
  "(.*)", index = "([ATGC]*|NoIndex)", lane = "([0-9]*)", read = "([0-9]*)", num
  = "([0-9]*)"), strict_format_checking = FALSE, sample_prefix = "")
}
\arguments{
\item{x}{path to a fastq folder}

\item{ext}{extensions to look for. A regex to search for fastq files}

\item{format}{auto detect. Pattern specify pattern acceptable to split_fastq_names. If missing will detect hiseq and miseq}

\item{lst_patterns}{a list of variables to be extracted, along with their regex patterns [typically the default should work for most cases]}

\item{strict_format_checking}{If some file names do not follow the format properly die with error (TRUE), continue with warning (FALSE)}

\item{sample_prefix}{A prefix to add to all sample names, run, project etc.....}
}
\description{
Create a table with details on names of fastq files
}

