% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sheets.R
\name{merge_sheets}
\alias{merge_sheets}
\title{merge tables by row, and filter rows}
\usage{
merge_sheets(x, outfile, .filter = NA, ...)
}
\arguments{
\item{x}{a character vector of file to be merged}

\item{outfile}{path to the merged output file}

\item{.filter}{a filter string. EXPERIMENTAL}

\item{...}{other arguments supplied to params::read_sheet}
}
\description{
merge tables by row, and filter rows
}
\examples{
\dontrun{
df = merge_sheets(c("mutect.chr1.txt", "mutect.chr2.txt"), 
   outfile = "mutect.merged.txt", .filter = "judgement==KEEP")
   
 dim(df)
}

}

