% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutect.R
\name{mutect}
\alias{mutect}
\title{A wrapper around somatic mutation caller MuTect}
\usage{
mutect(tumor_bam, normal_bam, samplename = opts_flow$get("samplename"),
  out_prefix, is_merged = TRUE, split_by_chr = TRUE,
  java_exe = opts_flow$get("java_exe"),
  java_tmp = opts_flow$get("java_tmp"),
  mutect_jar = opts_flow$get("mutect_jar"),
  cpu_mutect = opts_flow$get("cpu_mutect"),
  mem_mutect = opts_flow$get("java_mem"),
  ref_fasta = opts_flow$get("ref_fasta"),
  mutect_opts = opts_flow$get("mutect_opts"), execute_cmds = FALSE)
}
\arguments{
\item{tumor_bam}{path to a tumor bam file}

\item{normal_bam}{path to a normal bam file}

\item{samplename}{name of the sample, to be added to the flowmat}

\item{out_prefix}{output file name [optional].
By default, this is created using names of tumor and normal bam files.}

\item{is_merged}{specify if the input bam is already split by chromosomes (FALSE) 
or is a merged file with all chromosome (TRUE). [FALSE]}

\item{split_by_chr}{fork running mutect by chromosome to make it faster [TRUE].
Turning is option OFF is not fully supported.}

\item{java_exe}{path to java's executable [java]}

\item{java_tmp}{path to a temp folder to be used by java}

\item{mutect_jar}{path to mutect's jar file}

\item{cpu_mutect}{integer specifying number of thread to be used per mutect fork}

\item{mem_mutect}{amount of memory to be used by mutect [-Xmx12g]}

\item{ref_fasta}{path to the reference genome in fasta format}

\item{mutect_opts}{additional arguments passed onto mutect}
}
\value{
The function returns a flowmat with all the commands. 
The final file is called \code{'out_prefix'_merged.mutect.tsv}.
}
\description{
This generates a set of commandlines, per chromosome
}
\examples{
\dontrun{

x = "tumor.bam"
y = "normal.bam"

out = mutect(x, y, samplename="tumor_normal", is_merged = TRUE)

}
}

