% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_set.R
\name{paired_bam_set}
\alias{paired_bam_set}
\title{Title}
\usage{
paired_bam_set(tumor_bam, normal_bam, out_prefix, is_merged, split_by_chr,
  ref_fasta = opts_flow$get("ref_fasta"))
}
\arguments{
\item{tumor_bam}{}

\item{normal_bam}{}

\item{out_prefix}{}

\item{is_merged}{}

\item{split_by_chr}{}
}
\description{
Title
}
\details{
\code{
# single: tumor/normal
  |__: can create a out-prefix

multiple: tumor/normal
  |__: assumed one for each chr.
  |__: length   -  of both files should be same
  |__: sequence -  of files and those of chrs, is assumed to be the same
  |__: length of chr and tumor normal files should be the same
  | 
  |__: DICT is available. 
  |__: length and sequence of dict, is assumed (lexicographically sorted)
  |__: 
  |__: 
  |__: 
create a definitive out prefix
samplename, may not always be unique, using bams is better
determine output file name, if out_prefix is not provided
}
}

