% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{samtools_sort_index}
\alias{samtools_sort_index}
\title{set of wrapper functions around samtools}
\usage{
samtools_sort_index(x, outbam, samtools_exe)
}
\arguments{
\item{x}{input sam/bam file to be sorted and indexed}
}

