% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_fastqs.R
\name{split_names_fastq}
\alias{split_names_fastq}
\title{split.names.fastq}
\usage{
split_names_fastq(files,
  format = "$samplename$_$index$_L00$lane$_R$read$_$num$.fastq.gz")
}
\arguments{
\item{files}{a character vector of filenames, to be parsed}

\item{format}{the regex type format to be used to extract information from filenames. See default format as an example.}
}
\description{
Given a format split the files provided. Tried and tested on fastq files
}

