% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_fastqs.R
\name{split_names_fastq2}
\alias{split_names_fastq2}
\title{split_names_fastq2}
\usage{
split_names_fastq2(x,
  format = "{{samplename}}_{{index}}_L00{{lane}}_R{{read}}_{{num}}.fastq.gz",
  lst_patterns = list(samplename = "(.*)", index = "([ATGC]*|NoIndex)", lane =
  "([0-9]*)", read = "([0-9]*)", num = "([0-9]*)"),
  strict_format_checking = FALSE)
}
\arguments{
\item{x}{a fastq file}

\item{format}{naming format for the file}

\item{strict_format_checking}{TRUE (error)/FALSE(warning)}
}
\description{
split_names_fastq2
}

