% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_fastqs.R
\name{check_fastq_sheet}
\alias{check_fastq_sheet}
\title{check_fastq_sheet}
\usage{
check_fastq_sheet(x, id_column = "samplename", file_column = "file",
  read_column = "read")
}
\arguments{
\item{x}{a data.frame with details regarding fastq files}
}
\description{
Checks the nuber of samples, and that number of files in each sample should be
same for each read (if paired end)
}

