% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sheets.R
\name{merge_sheets}
\alias{merge_sheets}
\title{Title}
\usage{
merge_sheets(x, outfile, .filter = NA, ...)
}
\arguments{
\item{outfile}{path to the merged output file}

\item{fl}{a vector of files to be merged}
}
\description{
Title
}

